refactor(tests): Split medium-priority test files to comply with 300-line limit
Third batch of test refactoring: tackled medium-sized store and narrative
generator test files (400-500 lines) using established patterns.

**Store Tests Refactored:**
- navigationStore.test.ts (480 lines) → 3 focused files + helpers
  - routing.test.ts: path management, history, breadcrumbs
  - state.test.ts: modals, flow, preferences
  - persistence.test.ts: hydration, storage error handling

- worldStore.test.ts (448 lines) → 3 focused files + helpers
  - crud.test.ts: create, update, delete operations
  - attributes.test.ts: attribute management
  - skills.test.ts: skill management and settings

- npcStore.test.ts (395 lines) → 3 focused files + helpers
  - crud.test.ts: CRUD operations and retrieval
  - validation.test.ts: error handling and validation
  - world.test.ts: world management and state

**Narrative Generator Tests Refactored:**
- narrativeGenerator.inventory.test.ts (441 lines) → 3 focused files + helpers
  - access.test.ts: inventory data access
  - prioritization.test.ts: item prioritization logic
  - integration.test.ts: generation method integration

- narrativeGenerator.skillBasedChoices.test.ts (432 lines) → 3 focused files + helpers
  - skillAcknowledgment.test.ts: skill success/failure acknowledgment
  - skillChoices.test.ts: choice generation with skill requirements
  - skillContext.test.ts: skill context integration

**Changes:**
- Created test helper files for shared mock utilities and data factories
- Maintained test coverage and functionality (all tests passing)
- Applied consistent naming patterns across all split files
- Extracted reusable test data builders to reduce duplication

**Testing:**
All refactored test suites verified passing:
- navigationStore: 28 tests passing
- worldStore: 19 tests passing
- npcStore: 23 tests passing
- narrativeGenerator.inventory: 7 tests passing
- narrativeGenerator.skill*: 6 tests passing

Closes #830
