--- src/lib/ai/__tests__/narrativeGenerator.skillBasedChoices.test.ts
@@ -1,433 +1,0 @@
-/**
- * @fileoverview Tests for skill-based narrative choice generation
- * 
- * This test suite verifies that the narrative generator correctly creates
- * choices with skill requirements and acknowledges skill usage in follow-up narratives.
- * 
- * MVP Focus: Tests core functionality without over-engineering
- */
-
-import { NarrativeGenerator } from '../narrativeGenerator';
-import { AIClient } from '../types';
-import { NarrativeContext } from '@/types/narrative.types';
-import { World } from '@/types/world.types';
-import { Character } from '@/types/character.types';
-import { getTimestamp } from '@/lib/utils/timestamp';
-
-// Mock dependencies
-const mockAIClient: jest.Mocked<AIClient> = {
-  generateContent: jest.fn()
-};
-
-// Mock world store with skills
-const mockWorldWithSkills: World = {
-  id: 'skill-world',
-  name: 'Fantasy Realm',
-  description: 'A world with magic and combat',
-  genre: 'fantasy',
-  skills: [
-    {
-      id: 'athletics',
-      name: 'Athletics',
-      description: 'Physical prowess and endurance',
-      attributeIds: ['strength'],
-      worldId: 'skill-world',
-      difficulty: 'medium' as const,
-      baseValue: 0,
-      minValue: 0,
-      maxValue: 10
-    },
-    {
-      id: 'magic',
-      name: 'Magic',
-      description: 'Arcane knowledge and spellcasting',
-      attributeIds: ['intelligence'],
-      worldId: 'skill-world',
-      difficulty: 'hard' as const,
-      baseValue: 0,
-      minValue: 0,
-      maxValue: 10
-    },
-    {
-      id: 'stealth',
-      name: 'Stealth',
-      description: 'Moving unseen and unheard',
-      attributeIds: ['dexterity'],
-      worldId: 'skill-world',
-      difficulty: 'easy' as const,
-      baseValue: 0,
-      minValue: 0,
-      maxValue: 10
-    }
-  ],
-  attributes: [],
-  settings: {
-    maxAttributes: 6,
-    maxSkills: 12,
-    attributePointPool: 27,
-    skillPointPool: 40
-  },
-  createdAt: getTimestamp(),
-  updatedAt: getTimestamp()
-};
-
-// Mock character store
-const mockCharacterWithSkills: Character = {
-  id: 'char-1',
-  worldId: 'skill-world',
-  name: 'Test Hero',
-  description: 'A skilled adventurer ready for any challenge',
-  background: {
-    history: 'Skilled adventurer with years of experience',
-    personality: 'Brave and resourceful',
-    goals: ['Master all skills'],
-    fears: ['Failure'],
-    relationships: []
-  },
-  attributes: [
-    { attributeId: 'strength', value: 16 },
-    { attributeId: 'intelligence', value: 14 },
-    { attributeId: 'dexterity', value: 12 }
-  ],
-  skills: [
-    { skillId: 'athletics', level: 6, experience: 100, isActive: true },
-    { skillId: 'magic', level: 4, experience: 80, isActive: true },
-    { skillId: 'stealth', level: 3, experience: 60, isActive: true }
-  ],
-  inventory: {
-    characterId: 'char-1',
-    items: [],
-    capacity: 100,
-    categories: [],
-    itemOrder: []
-  },
-  status: {
-    health: 100,
-    maxHealth: 100,
-    conditions: []
-  },
-  createdAt: getTimestamp(),
-  updatedAt: getTimestamp()
-};
-
-jest.mock('@/state/worldStore', () => ({
-  useWorldStore: {
-    getState: () => ({
-      worlds: {
-        'skill-world': mockWorldWithSkills
-      },
-      currentWorldId: 'skill-world'
-    })
-  }
-}));
-
-jest.mock('@/state/characterStore', () => ({
-  useCharacterStore: {
-    getState: () => ({
-      characters: {
-        'char-1': mockCharacterWithSkills
-      }
-    })
-  }
-}));
-
-// Mock template manager
-jest.mock('../../promptTemplates/narrativeTemplateManager', () => ({
-  narrativeTemplateManager: {
-    getTemplate: jest.fn().mockImplementation((key: string) => {
-      if (key === 'narrative/scene') {
-        return jest.fn().mockImplementation((context) => {
-          const characterSkillContext = context.characterSkillContext || '';
-          return `Generate a scene...${characterSkillContext}`;
-        });
-      }
-      if (key === 'narrative/skillAcknowledgment') {
-        return jest.fn().mockReturnValue('Generate skill acknowledgment...');
-      }
-      return jest.fn().mockReturnValue('Template not found');
-    })
-  }
-}));
-
-describe('NarrativeGenerator - Skill-Based Choices', () => {
-  let narrativeGenerator: NarrativeGenerator;
-  
-  beforeEach(() => {
-    jest.clearAllMocks();
-    narrativeGenerator = new NarrativeGenerator(mockAIClient);
-  });
-
-  describe('generateSegment with skill acknowledgment', () => {
-    it('should generate narrative that acknowledges successful skill usage', async () => {
-      // Mock AI response that acknowledges skill usage
-      const mockResponse = {
-        content: JSON.stringify({
-          content: "Your athletic training pays off as you leap across the chasm with ease. The crowd cheers as you land gracefully on the other side, your strength clearly evident in the powerful jump.",
-          metadata: {
-            location: "Canyon Bridge",
-            mood: "triumphant",
-            tags: ["skill-success", "athletics", "crowd-reaction"],
-            characterIds: ["char-1"]
-          }
-        }),
-        finishReason: 'stop' as const,
-        promptTokens: 75,
-        completionTokens: 75
-      };
-      mockAIClient.generateContent.mockResolvedValue(mockResponse);
-
-      const narrativeContext: NarrativeContext = {
-        worldId: 'skill-world',
-        currentSceneId: 'scene-1',
-        characterIds: ['char-1'],
-        sessionId: 'session-1',
-        previousSegments: [],
-        currentTags: ['athletics-used'],
-        currentLocation: 'Canyon Bridge',
-        currentSituation: 'After successfully using athletics to leap the chasm'
-      };
-
-      const result = await narrativeGenerator.generateSegment({
-        worldId: 'skill-world',
-        sessionId: 'session-1',
-        characterIds: ['char-1'],
-        narrativeContext
-      });
-
-      expect(result.content).toContain('athletic training pays off');
-      expect(result.content).toContain('strength clearly evident');
-      expect(result.metadata.tags).toContain('skill-success');
-      expect(result.metadata.location).toBe('Canyon Bridge');
-    });
-
-    it('should generate narrative that handles failed skill checks', async () => {
-      // Mock AI response for skill failure
-      const mockResponse = {
-        content: JSON.stringify({
-          content: "Despite your best efforts, the spell fizzles out before completion. Your magical energy feels drained, and you realize you need more practice with this level of enchantment.",
-          metadata: {
-            location: "Magic Tower",
-            mood: "tense",
-            tags: ["skill-failure", "magic", "consequence"],
-            characterIds: ["char-1"]
-          }
-        }),
-        finishReason: 'stop' as const,
-        promptTokens: 70,
-        completionTokens: 70
-      };
-      mockAIClient.generateContent.mockResolvedValue(mockResponse);
-
-      const narrativeContext: NarrativeContext = {
-        worldId: 'skill-world',
-        currentSceneId: 'scene-2',
-        characterIds: ['char-1'],
-        sessionId: 'session-1',
-        previousSegments: [],
-        currentTags: ['magic-failed'],
-        currentLocation: 'Magic Tower',
-        currentSituation: 'After failing a magic skill check'
-      };
-
-      const result = await narrativeGenerator.generateSegment({
-        worldId: 'skill-world',
-        sessionId: 'session-1',
-        characterIds: ['char-1'],
-        narrativeContext
-      });
-
-      expect(result.content).toContain('spell fizzles out');
-      expect(result.content).toContain('need more practice');
-      expect(result.metadata.tags).toContain('skill-failure');
-      expect(result.metadata.mood).toBe('tense');
-    });
-  });
-
-  describe('choice generation with skill requirements', () => {
-    it('should pass character skills to choice generator', async () => {
-      // Mock the choice generator to spy on the context passed to it
-      const mockChoiceGenerator = {
-        generateChoices: jest.fn().mockResolvedValue({
-          id: 'decision-with-skills',
-          prompt: 'What approach will you take?',
-          options: [
-            {
-              id: 'opt-1',
-              text: 'Climb the wall',
-              requirements: [{ type: 'skill' as const, targetId: 'athletics', operator: 'gte' as const, value: 5 }],
-              hint: 'Requires athletic ability'
-            },
-            {
-              id: 'opt-2', 
-              text: 'Cast a levitation spell',
-              requirements: [{ type: 'skill' as const, targetId: 'magic', operator: 'gte' as const, value: 4 }],
-              hint: 'Requires magical knowledge'
-            },
-            {
-              id: 'opt-3',
-              text: 'Look for another way',
-              alignment: 'neutral' as const
-            }
-          ]
-        })
-      };
-
-      // Replace the choice generator in the narrative generator
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (narrativeGenerator as any).choiceGenerator = mockChoiceGenerator;
-
-      const narrativeContext: NarrativeContext = {
-        worldId: 'skill-world',
-        currentSceneId: 'scene-3',
-        characterIds: ['char-1'],
-        sessionId: 'session-1',
-        previousSegments: [],
-        currentTags: ['obstacle'],
-        currentLocation: 'High Wall',
-        currentSituation: 'Facing a tall obstacle'
-      };
-
-      await narrativeGenerator.generatePlayerChoices(
-        'skill-world',
-        narrativeContext,
-        ['char-1']
-      );
-
-      // Verify the choice generator was called with correct parameters
-      expect(mockChoiceGenerator.generateChoices).toHaveBeenCalledWith({
-        worldId: 'skill-world',
-        narrativeContext,
-        characterIds: ['char-1'],
-        minOptions: 3,
-        maxOptions: 4,
-        useAlignedChoices: false
-      });
-
-      // Verify the choice generator was called - the choices themselves are tested in choiceGenerator tests
-    });
-  });
-
-  describe('skill context integration', () => {
-    it('should include character skill information in narrative context', async () => {
-      const mockResponse = {
-        content: "You assess your options, drawing on your athletic prowess and magical knowledge.",
-        finishReason: 'stop' as const,
-        promptTokens: 50,
-        completionTokens: 50
-      };
-      mockAIClient.generateContent.mockResolvedValue(mockResponse);
-
-      await narrativeGenerator.generateSegment({
-        worldId: 'skill-world',
-        sessionId: 'session-1',
-        characterIds: ['char-1'],
-        narrativeContext: {
-          worldId: 'skill-world',
-          currentSceneId: 'scene-4',
-          characterIds: ['char-1'],
-          sessionId: 'session-1',
-          previousSegments: [],
-          currentTags: []
-        }
-      });
-
-      // Verify that the AI client was called with a prompt
-      expect(mockAIClient.generateContent).toHaveBeenCalled();
-      const calledPrompt = mockAIClient.generateContent.mock.calls[0][0];
-      
-      // The prompt should include character skill context
-      expect(typeof calledPrompt).toBe('string');
-      expect(calledPrompt.length).toBeGreaterThan(0);
-    });
-  });
-
-  describe('MVP skill acknowledgment requirements', () => {
-    it('should generate different narrative responses based on skill success/failure context', async () => {
-      // Test successful skill usage narrative
-      const successResponse = {
-        content: "Your expertise shines through as you execute the maneuver perfectly.",
-        finishReason: 'stop' as const,
-        promptTokens: 45,
-        completionTokens: 45
-      };
-      mockAIClient.generateContent.mockResolvedValueOnce(successResponse);
-
-      const successContext: NarrativeContext = {
-        worldId: 'skill-world',
-        currentSceneId: 'scene-success',
-        characterIds: ['char-1'],
-        sessionId: 'session-1',
-        previousSegments: [],
-        currentTags: ['skill-success', 'athletics'],
-        currentSituation: 'Successfully completed athletic challenge'
-      };
-
-      const successResult = await narrativeGenerator.generateSegment({
-        worldId: 'skill-world',
-        sessionId: 'session-1',
-        characterIds: ['char-1'],
-        narrativeContext: successContext
-      });
-
-      expect(successResult.content).toContain('expertise');
-
-      // Test failed skill usage narrative
-      const failureResponse = {
-        content: "You struggle with the technique, clearly needing more practice.",
-        finishReason: 'stop' as const,
-        promptTokens: 42,
-        completionTokens: 43
-      };
-      mockAIClient.generateContent.mockResolvedValueOnce(failureResponse);
-
-      const failureContext: NarrativeContext = {
-        worldId: 'skill-world',
-        currentSceneId: 'scene-failure',
-        characterIds: ['char-1'],
-        sessionId: 'session-1',
-        previousSegments: [],
-        currentTags: ['skill-failure', 'magic'],
-        currentSituation: 'Failed magical attempt'
-      };
-
-      const failureResult = await narrativeGenerator.generateSegment({
-        worldId: 'skill-world',
-        sessionId: 'session-1',
-        characterIds: ['char-1'],
-        narrativeContext: failureContext
-      });
-
-      expect(failureResult.content).toContain('struggle');
-    });
-
-    it('should handle worlds without skills gracefully', async () => {
-      // Reset mocks and use a simple successful response
-      mockAIClient.generateContent.mockResolvedValue({
-        content: "You proceed with your adventure in this simple world.",
-        finishReason: 'stop' as const,
-        promptTokens: 35,
-        completionTokens: 35
-      });
-
-      // Use the existing mock world from setup but don't override skills
-      // The existing system should handle empty skills gracefully
-      const result = await narrativeGenerator.generateSegment({
-        worldId: 'skill-world', // Use existing mock world
-        sessionId: 'session-1',
-        characterIds: ['char-1'],
-        narrativeContext: {
-          worldId: 'skill-world',
-          currentSceneId: 'scene-simple',
-          characterIds: ['char-1'],
-          sessionId: 'session-1',
-          previousSegments: [],
-          currentTags: []
-        }
-      });
-
-      expect(result).toBeDefined();
-      expect(result.content).toBeTruthy();
-      expect(result.content).toBe("You proceed with your adventure in this simple world.");
-    });
-  });
-});
