+++ src/state/__tests__/npcStore.testHelpers.ts
@@ -1,0 +1,35 @@
+/**
+ * Test helpers for npcStore tests
+ * Provides reusable NPC data factories
+ */
+
+/**
+ * Creates basic NPC test data
+ */
+export function createTestNPCData(overrides?: Partial<{
+  worldId: string;
+  name: string;
+  description: string;
+  avatarUrl?: string;
+}>) {
+  return {
+    worldId: 'world-123',
+    name: 'Test NPC',
+    description: 'A test NPC for testing',
+    ...overrides
+  };
+}
+
+/**
+ * Resets the NPC store to initial state
+ */
+export function resetNPCStore(useNPCStore: { setState: (state: unknown) => void }) {
+  useNPCStore.setState({
+    npcs: {},
+    entities: {},
+    worldNpcs: {},
+    currentEntityId: null,
+    error: null,
+    loading: false,
+  });
+}
